+++ examples/java/azure/src/main/java/com/mycompany/app/Main.java
@@ -17,7 +17,9 @@ public class Main extends TerraformStack {
     public Main(final Construct scope, final String id) {
         super(scope, id);
 
-        AzurermProvider.Builder.create(this, "AzureRm").features(new ArrayList<AzurermProviderFeatures>()).build();
+        AzurermProvider.Builder.create(this, "AzureRm")
+            .features(Arrays.asList(AzurermProviderFeatures.builder().build()))
+            .build();
 
         VirtualNetwork.Builder.create(this, "TfVnet")
             .location("uksouth")